# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "project-forge API",
    "version": "1.0.0",
    "description": "API for creating projects using planforge + scaffoldkit toolchain with automatic GitHub repository creation.",
    "contact": {
      "name": "project-forge",
      "url": "https://project-forge.opentriologue.ai"
    }
  },
  "servers": [
    {
      "url": "https://project-forge.opentriologue.ai",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/api/v1/projects": {
      "get": {
        "summary": "List projects",
        "description": "List projects (usage log entries) created by the authenticated user's API token, most recent first.",
        "operationId": "listProjects",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Max number of results (capped at 200).",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Number of results to skip.",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "includeDeleted",
            "in": "query",
            "required": false,
            "description": "Include soft-deleted projects in the results.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects listed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProjectsResponse"
                },
                "example": {
                  "ok": true,
                  "projects": [
                    {
                      "id": "cljk3x9q10000abcd1234efgh",
                      "repoUrl": "https://github.com/username/my-awesome-app",
                      "projectName": "my-awesome-app",
                      "createdAt": "2026-08-01T12:00:00.000Z"
                    }
                  ],
                  "total": 1,
                  "limit": 50,
                  "offset": 0
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingKey": {
                    "summary": "Missing API key",
                    "value": {
                      "ok": false,
                      "error": "Missing X-API-Key header"
                    }
                  },
                  "invalidKey": {
                    "summary": "Invalid API key",
                    "value": {
                      "ok": false,
                      "error": "Invalid or revoked API token"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Soft-delete a project",
        "description": "Mark a project (usage log entry) owned by the authenticated user's token as deleted.",
        "operationId": "deleteProject",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "The project (usage log entry) id to delete.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing id parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingId": {
                    "summary": "Missing id parameter",
                    "value": {
                      "ok": false,
                      "error": "Missing id parameter"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingKey": {
                    "summary": "Missing API key",
                    "value": {
                      "ok": false,
                      "error": "Missing X-API-Key header"
                    }
                  },
                  "invalidKey": {
                    "summary": "Invalid API key",
                    "value": {
                      "ok": false,
                      "error": "Invalid or revoked API token"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "notFound": {
                    "summary": "Project not found",
                    "value": {
                      "ok": false,
                      "error": "Project not found"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Project already deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "alreadyDeleted": {
                    "summary": "Already deleted",
                    "value": {
                      "ok": false,
                      "error": "Already deleted"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new project",
        "description": "Generate a new project using planforge and scaffoldkit, create a GitHub repository, and push the initial commit. Rate limit: 10 project creations per user per day.",
        "operationId": "createProject",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRequest"
              },
              "examples": {
                "basic": {
                  "summary": "Basic project",
                  "value": {
                    "projectName": "my-awesome-app",
                    "summary": "A web application for managing tasks",
                    "features": [
                      "User authentication",
                      "Task CRUD operations",
                      "Real-time updates"
                    ],
                    "constraints": [
                      "Use TypeScript",
                      "Deploy on Vercel"
                    ]
                  }
                },
                "minimal": {
                  "summary": "Minimal project",
                  "value": {
                    "projectName": "simple-cli",
                    "summary": "A command-line tool for file processing"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectResponse"
                },
                "example": {
                  "ok": true,
                  "result": {
                    "repoUrl": "https://github.com/username/my-awesome-app",
                    "cloneUrl": "https://github.com/username/my-awesome-app.git",
                    "projectName": "my-awesome-app"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields or GitHub PAT not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingFields": {
                    "summary": "Missing required fields",
                    "value": {
                      "ok": false,
                      "error": "Missing required fields: projectName, summary"
                    }
                  },
                  "missingPat": {
                    "summary": "GitHub PAT not configured",
                    "value": {
                      "ok": false,
                      "error": "GitHub PAT not configured. Please add it in your dashboard."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingKey": {
                    "summary": "Missing API key",
                    "value": {
                      "ok": false,
                      "error": "Missing X-API-Key header"
                    }
                  },
                  "invalidKey": {
                    "summary": "Invalid API key",
                    "value": {
                      "ok": false,
                      "error": "Invalid or revoked API token"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "ok": false,
                  "error": "Rate limit exceeded",
                  "details": "Maximum 10 projects per day. Used: 10/10"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - project generation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "ok": false,
                  "error": "Project generation failed",
                  "details": "planforge execution failed: timeout"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API token generated from your dashboard. Format: pf_*"
      }
    },
    "schemas": {
      "ProjectRequest": {
        "type": "object",
        "required": [
          "projectName",
          "summary"
        ],
        "properties": {
          "projectName": {
            "type": "string",
            "description": "Name of the project (will be used as GitHub repository name). This schema is shared by POST /api/v1/projects and POST /api/v1/generate; the pattern reflects generate's real validator (lib/v1-shared.ts PROJECT_NAME_RE). POST /api/v1/projects does not enforce this pattern itself -- it only rejects an empty projectName -- so a wider valid name there may fail if later passed through generate.",
            "example": "my-awesome-app",
            "pattern": "^[a-zA-Z0-9._-]{1,100}$"
          },
          "summary": {
            "type": "string",
            "description": "Brief description of what the project does",
            "example": "A web application for managing tasks"
          },
          "features": {
            "type": "array",
            "description": "List of core features the project should have",
            "items": {
              "type": "string"
            },
            "example": [
              "User authentication",
              "Task CRUD",
              "Real-time updates"
            ]
          },
          "constraints": {
            "type": "array",
            "description": "Technical constraints or requirements",
            "items": {
              "type": "string"
            },
            "example": [
              "Use TypeScript",
              "Deploy on Vercel"
            ]
          },
          "targetUsers": {
            "type": "array",
            "description": "Target user groups",
            "items": {
              "type": "string"
            },
            "example": [
              "developers",
              "project managers"
            ],
            "default": [
              "developers"
            ]
          }
        }
      },
      "CreateProjectResponse": {
        "type": "object",
        "required": [
          "ok",
          "result"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "result": {
            "$ref": "#/components/schemas/PublishResult"
          }
        }
      },
      "PublishResult": {
        "type": "object",
        "description": "The v1 API's publish result shape (app/api/v1/publish/route.ts). Same name as lib/types.ts PublishResult, but that one is shaped { repoUrl, cloneUrl, defaultBranch } for the legacy UI route -- the two are not interchangeable.",
        "properties": {
          "repoUrl": {
            "type": "string",
            "description": "GitHub repository URL",
            "example": "https://github.com/username/my-awesome-app"
          },
          "cloneUrl": {
            "type": "string",
            "description": "Git clone URL (HTTPS)",
            "example": "https://github.com/username/my-awesome-app.git"
          },
          "projectName": {
            "type": "string",
            "description": "Name of the created project",
            "example": "my-awesome-app"
          }
        }
      },
...
      "ListProjectsResponse": {
        "type": "object",
        "required": [
          "ok",
          "projects",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectListItem"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total matching projects (ignores limit/offset).",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 50
          },
          "offset": {
            "type": "integer",
            "example": 0
          }
        }
      },
      "ProjectListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cljk3x9q10000abcd1234efgh"
          },
          "repoUrl": {
            "type": "string",
            "example": "https://github.com/username/my-awesome-app"
          },
          "projectName": {
            "type": "string",
            "description": "Derived from the trailing path segment of repoUrl.",
            "example": "my-awesome-app"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OkResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error shape for app/api/** (lib/types.ts ErrorResponse). `ok` is always `false` on an error response.",
        "required": [
          "ok",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid or revoked API token"
          },
          "details": {
            "type": "string",
            "description": "Additional error details (optional)",
            "example": "Token was revoked on 2026-03-25"
          }
        }
      }
    }
  }
}